[Refactor] Modularize answers.py runner and fix display bugs
- Split 700-line answers.py into modular components under solutions/euler/util/run/
  - performance.py: Performance categorization and threshold checking
  - runner.py: Solution execution with timeout handling
  - display.py: Rich table output and summary formatting
  - cli.py: Command-line argument parsing
- Fix status display bug: state column now shows actual status (pass/failed/timeout/error)
- Fix table width calculation to properly account for Rich borders
- Update notebooks (p012, p0103, p0106) with proper title/link format per preferences
